Add stencil test, scissor test & some cleaning.

+++ Aka/OS/Image.h
@@ -15,10 +15,10 @@
 namespace aka {
 
 struct Rect {
-	float x;
-	float y;
-	float w;
-	float h;
+	int32_t x;
+	int32_t y;
+	uint32_t w;
+	uint32_t h;
 };
 
 struct Image
